Updating server disks fix
A fix for the server update command that allows the disks in the
server to be updated. Previously you had to specify both disks to
keep and disks to add. This isn't always the case and so the changes
makes it possible to either keep disks (and update their size) or add
new disks or both.

Resolves: #69

Signed-off-by: Richard Case <richard.case@outlook.com>