Various fixes - cross dc, get server and server create (#82)
* Added get baremetal capabilities call
Enhancement to the CLI that introduces a new 'get-baremetal-capabilities' command on the
data-center resource. This calls the following API:
https://www.ctl.io/api-docs/v2/#data-centers-get-data-center-bare-metal-capabilities

* Fixes for cross dc and get server details

This contains a fix for issue #64 where getting the cross data centre
policies didn't filter as expected when specifying a destination account
alias. Also, fixed issue #67 where the IsManaged and IsManagedBackup
wasn't being returned when getting server details.

Resolves: #64, #67

* Updating server disks fix

A fix for the server update command that allows the disks in the
server to be updated. Previously you had to specify both disks to
keep and disks to add. This isn't always the case and so the changes
makes it possible to either keep disks (and update their size) or add
new disks or both.

Resolves: #69

* Bare metal create parameter updates

The server create command has been updated so that the  parameters such as
CPU, MemoryGB, template, Source Server Id are only required for standard and
hyperscale servers. The descriptions for 'configuration-id' and 'os-type'
have been updated to indicate that the expected values can be obtained
by calling data-center get-baremetal-capabilities.

Resolves: #75

Signed-off-by: Richard Case <richard.case@outlook.com>